% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{add_line_num}
\alias{add_line_num}
\title{Add line numbers for a head patch}
\usage{
add_line_num(head_df)
}
\arguments{
\item{head_df}{(tibble) with 4 columns: \code{raw}, \code{hunk}, \code{head}, and
\code{context}}
}
\value{
a tibble with an extra integer column, \code{line} representing
the line number in "head"
}
\description{
Add line numbers for a head patch
}
\keyword{internal}
