% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{build_badge_url}
\alias{build_badge_url}
\title{Build the badge URL}
\usage{
build_badge_url(percentage = NULL)
}
\arguments{
\item{percentage}{(numeric)}
}
\value{
a character scalar representing the URL to the SVG badge
}
\description{
Badge is generated with shields.io
}
\examples{
build_badge_url(15.43)
}
