% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment.R
\name{comm_delete}
\alias{comm_delete}
\title{Delete a comment}
\usage{
comm_delete(repo, comment_id)
}
\arguments{
\item{repo}{(character) the repository name in the GitHub format
(\code{"OWNER/REPO"}).}

\item{comment_id}{(numeric) the ID of the issue comment to delete.}
}
\value{
a \code{gh_response} object
}
\description{
Thin wrapper for making a \code{DELETE} request to the issue comments endpoint
of the GitHub API.
}
\examples{
\dontrun{
comm_delete("<owner>/<repo>", 4553)
}
}
\keyword{internal}
