% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment.R
\name{compose_comment}
\alias{compose_comment}
\title{Compose a coverage GitHub comment}
\usage{
compose_comment(head_coverage, base_coverage, owner, repo, pr_number)
}
\arguments{
\item{head_coverage}{(coverage) active / current branch (\code{HEAD}) coverage.
The output of \code{\link[covr:package_coverage]{covr::package_coverage()}} on the branch.}

\item{base_coverage}{(coverage) base / target branch coverage (coverage for
the branch merging into). The output of \code{\link[covr:package_coverage]{covr::package_coverage()}} on the
branch.}

\item{owner}{(character scalar) repo owner}

\item{repo}{(character scalar) repo name}

\item{pr_number}{(integerish) pull request number}
}
\value{
a character scalar with the content of the GitHub comment
}
\description{
Compose a coverage GitHub comment
}
\examples{
\dontrun{
head_coverage <- covr::package_coverage()
system2("git", c("checkout", "main"))
base_coverage <- covr::package_coverage()

compose_comment(
  head_coverage = head_coverage,
  base_coverage = base_coverage,
  owner = "dragosmg",
  repo = "covr2mddemo",
  pr_number = 3
)
}
}
