% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{compose_file_cov_details}
\alias{compose_file_cov_details}
\title{Compose coverage details}
\usage{
compose_file_cov_details(
  head_coverage,
  base_coverage,
  changed_files,
  align = "rrrrc"
)
}
\arguments{
\item{head_coverage}{(coverage) active / current branch (\code{HEAD}) coverage.
The output of \code{\link[covr:package_coverage]{covr::package_coverage()}} on the branch.}

\item{base_coverage}{(coverage) base / target branch coverage (coverage for
the branch merging into). The output of \code{\link[covr:package_coverage]{covr::package_coverage()}} on the
branch.}

\item{changed_files}{(character) names of files changed by the PR. Usually
the output of \code{\link[=get_changed_files]{get_changed_files()}}.}
}
\value{
a markdown table of changes in coverage at file level between the
head and base branches.
}
\description{
Compose coverage details
}
\examples{
\dontrun{
changed_files <- get_changed_files(
  repo = "dragosmg/covr2mddemo",
  pr_number = 2
)

file_coverage_details <- compose_file_cov_details(
  head_coverage = head_coverage,
  base_coverage = base_coverage,
  changed_files = changed_files
)
}
}
\keyword{internal}
