% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_coverage.R
\name{compose_line_coverage_summary}
\alias{compose_line_coverage_summary}
\title{Compose the line coverage summary}
\usage{
compose_line_coverage_summary(line_cov_delta, target = 80, our_target = FALSE)
}
\arguments{
\item{line_cov_delta}{a \code{tibble} the output of \code{\link[=diff_line_coverage]{diff_line_coverage()}}}

\item{target}{(numeric) the target coverage for the diff. Defaults to 80, but
\code{compose_comment()} uses the total coverage for base.}
}
\value{
a glue string, a sentence summarising the diff coverage.
}
\description{
Builds the high-level sentence summarising the line coverage of the patch.
}
\keyword{internal}
