% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest.R
\name{digest_coverage}
\alias{digest_coverage}
\title{Calculate file coverage}
\usage{
digest_coverage(x = covr::package_coverage())
}
\arguments{
\item{x}{\verb{<coverage>} object, defaults to \code{\link[covr:package_coverage]{covr::package_coverage()}}.}
}
\value{
a \code{tibble} with 2 columns (\code{File} and \code{Coverage}) summarising
testing coverage at file level.
}
\description{
Calculate file coverage
}
\examples{
\dontrun{
library(covr)

covr::package_coverage("myawesomepkg") |>
  digest_coverage()
}
}
