% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_coverage.R
\name{file_coverage}
\alias{file_coverage}
\title{Digest file coverage}
\usage{
file_coverage(coverage, call = rlang::caller_env())
}
\arguments{
\item{coverage}{\verb{<coverage>} object, defaults to \code{\link[covr:package_coverage]{covr::package_coverage()}}.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
a \code{tibble} with 2 columns (\code{File} and \code{Coverage}) summarising
testing coverage at file level.
}
\description{
Takes a \code{coverage} object (the output of \code{\link[covr:package_coverage]{covr::package_coverage()}}, extracts
the \code{"filecoverage"} component and transforms it into a \code{tibble}. It also
extracts the \code{"totalcoverage"} and adds it as the \code{"Overall"} row.
}
\examples{
\dontrun{
library(covr)

covr::package_coverage("myawesomepkg") |>
  file_coverage()
}
}
\keyword{internal}
