% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_coverage.R
\name{file_coverage_digest}
\alias{file_coverage_digest}
\title{Digest file coverage}
\usage{
file_coverage_digest(coverage, call = rlang::caller_env())
}
\arguments{
\item{coverage}{a \verb{<coverage>} object, the output of a call to
\code{\link[covr:package_coverage]{covr::package_coverage()}}.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
a \code{tibble} with 2 columns (\code{File} and \code{Coverage}) summarising
testing coverage at file level.
}
\description{
Takes a \code{coverage} object (the output of \code{\link[covr:package_coverage]{covr::package_coverage()}}, extracts
the \code{"filecoverage"} component and transforms it into a \code{tibble}.
Additionally, it extracts the \code{"totalcoverage"} element and adds it as the
\code{"Overall"} row.
}
\examples{
\dontrun{
library(covr)

coverage <- covr::package_coverage("myawesomepkg")
file_coverage_digest(coverage)
}
}
\keyword{internal}
