% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_details.R
\name{get_diff_text}
\alias{get_diff_text}
\title{Get the PR diff}
\usage{
get_diff_text(pr_details)
}
\arguments{
\item{pr_details}{a \code{pr_details} object.}
}
\value{
a named list where the names are file names and the content of each
element is the patch for the specific file.
}
\description{
Sends a GET request to the GitHub API and retrieves the full PR diff, which
is a (comparison) between base (the starting point for the comparison) and
head (the endpoint). The diff is then filtered to only include the "relevant
files".
}
\examples{
\dontrun{
pr_details <- get_pr_details("<owner>/<repo>", 2)

diff_text <- get_diff_text(pr_details)
}
}
\keyword{internal}
