% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment.R
\name{post_comment}
\alias{post_comment}
\title{Post comment}
\usage{
post_comment(body, repo, pr_number, update = TRUE)
}
\arguments{
\item{body}{(character scalar) the content of the body of the message.}

\item{repo}{(character) the repository name in the GitHub format
(\code{"OWNER/REPO"}).}

\item{pr_number}{(integer) the PR number}

\item{update}{(logical) update an existing comment or post a new one.
Defaults to \code{TRUE}.}
}
\value{
a \code{gh_response} object containing the API response
}
\description{
\code{post_comment()} first checks if a "known" \code{covr2gh} comment exists on the
target pull request. If it does, then it updates it, if it doesn't, then a
a new comment is posted.
}
\details{
Users can also choose to always post a new comment (this always deletes the
previous one).
}
\examples{
\dontrun{
post_comment(
  "this is amazing",
  repo = "<owner>/<repo>",
  pr_number = 3
)
}
}
