% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_coverage.R
\name{file_coverage}
\alias{file_coverage}
\title{Digest file coverage}
\usage{
file_coverage(x)
}
\arguments{
\item{x}{\verb{<coverage>} object, defaults to \code{\link[covr:package_coverage]{covr::package_coverage()}}.}
}
\value{
a \code{tibble} with 2 columns (\code{File} and \code{Coverage}) summarising
testing coverage at file level.
}
\description{
Takes a \code{coverage} object (the output of \code{\link[covr:package_coverage]{covr::package_coverage()}}, extracts
the \code{"filecoverage"} component and transforms it into a \code{tibble}. It also
extracts the \code{"totalcoverage"} and adds it as the \code{"Overall"} row.
}
\examples{
\dontrun{
library(covr)

covr::package_coverage("myawesomepkg") |>
  file_coverage()
}
}
\keyword{internal}
